% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_genes_by_cnv.R
\name{annotate_genes_by_cnv}
\alias{annotate_genes_by_cnv}
\title{Annotate genes by CNV status}
\usage{
annotate_genes_by_cnv(segs, genes = NULL, n.cores = 1)
}
\arguments{
\item{segs}{list. Output of run_segmentation.}

\item{genes}{GRanges object containing genes or transcripts. If \code{NULL},
the program calls \code{sesameData::sesameData_getTxnGRanges("hg38", merge2gene=TRUE)}.}

\item{n.cores}{integer. Specify how many cores to be used for computation (default: 1).}
}
\value{
list. An expanded list that includes 'genes' annotated with overlapping segments.
}
\description{
This function can take long time to complete for a large number of genes.
}
