% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_gene_cnv.R
\name{detect_gene_cnv}
\alias{detect_gene_cnv}
\title{Detect gene-level CNVs}
\usage{
detect_gene_cnv(
  segs,
  crs = NULL,
  genes,
  extend.bp = 5e+05,
  conf = 0.99,
  R = 100,
  blockLength = 5e+05,
  proportionLength = TRUE,
  use.denoised = FALSE
)
}
\arguments{
\item{crs}{list. Output of \code{tangent_normalization}. Necessary only when probe-binning was used (default: NULL).}

\item{genes}{GenomicRanges object (predefined gene list is available: \code{data(genes.hg38)})}

\item{extend.bp}{integer. Extend gene interval for \code{+/- extend.bp} (default: \code{500000L}).}

\item{conf}{numeric. Confidence level to calculate to determine the log2-threshold for
high-level alterations. Choose between \code{0.95} and \code{0.99}. Default to \code{0.99}.}

\item{R}{numeric. Parameter for the \code{bootRanges} function.
The number of bootstrap samples to generate. Default to \code{100}.}

\item{blockLength}{numeric. Parameter for the \code{bootRanges} function. The length
(in basepairs) of the blocks for segmented block bootstrapping. Default to \code{500000}.}

\item{proportionLength}{logical. From the \code{nullranges} package: for the segmented block
bootstrap, whether to use scaled block lengths, (scaling by the proportion of the segmentation
state out of the total genome length)}

\item{use.denoised}{logical. Use raw signal (use.denoised = FALSE) or denoised signal (use.denoised = TRUE).}

\item{segdata}{list. Output of \code{run_segmentation}.}
}
\value{
list. genes, tbl of gene-level result. amp/del_threshold, thresholds for amplification
and deletions determined by bootRanges.
}
\description{
Detect gene-level (or any local region) CNVs. Adopted from conumee2.0
(https://github.com/hovestadtlab/conumee2.git)
}
