% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genomic_coord.R
\name{get_genomic_coord}
\alias{get_genomic_coord}
\title{Transform chromosomal coordinates to genomic coordinates}
\usage{
get_genomic_coord(df, hg38.seqinfo, space.skip = 0.01)
}
\arguments{
\item{df}{A tbl.}

\item{hg38.seqinfo}{Use \code{data("hg38.seqinfo")}}

\item{space.skip}{Space between chromosomes for genome-wide plot (default: 0.01)}
}
\value{
list
}
\description{
Transform chromosomal coordinates to genomic coordinates
}
