% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probe_bins.R
\name{get_probe_bins}
\alias{get_probe_bins}
\title{Get probe bins}
\usage{
get_probe_bins(crs, bin.size = 1000L, hg38.seqinfo, centromere.hg38)
}
\arguments{
\item{crs}{list. Output of \code{tangent_normalization}.}

\item{bin.size}{integer. In base pairs (default: 1000L).}

\item{hg38.seqinfo}{Use \code{data("hg38.seqinfo")}.}

\item{centromere.hg38}{Use \code{data("centromere.hg38")}.}
}
\value{
tbl. seqnames, start, end, width, nProbes, chrArm and median log R ratio (lrr).
}
\description{
Get probe bins
}
