% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sex_info.R
\name{get_sex_info}
\alias{get_sex_info}
\title{get probe data for sex inference}
\usage{
get_sex_info(sdf, platform, cutoff = -3.5)
}
\arguments{
\item{sdf}{SigDF object.}

\item{platform}{character. HM450, EPIC, or EPICv2.}

\item{cutoff}{a cutoff value of log2(medianY/medianX) for sex inference (default: -3.5).}
}
\value{
a named list with pred.sex (predicted sex), medianX (median intensity of X probes), and
medianY (median intensity of Y probes).
}
\description{
get probe data for sex inference
}
