% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_copyRatio.R
\name{plot_copyRatio}
\alias{plot_copyRatio}
\title{Plot copy ratio data}
\usage{
plot_copyRatio(
  segs,
  crs = NULL,
  hg38.seqinfo,
  centromere.hg38,
  ymax = 1,
  space.skip = 0.01,
  loss.lrr = NULL,
  gain.lrr = NULL
)
}
\arguments{
\item{segs}{segment data (seqnames, start, end, mean.lrr, Z.lrr)}

\item{crs}{denoised copy ratio data (seqnames, start, end, lrr).}

\item{hg38.seqinfo}{Use \code{data("hg38.seqinfo")}.}

\item{centromere.hg38}{Use \code{data("centromere.hg38")}.}

\item{ymax}{Maximum copy ratio (default: 3).}

\item{space.skip}{Space between chromosomes (default: 0.01)}

\item{loss.lrr}{numeric. Cutoff value of lrr for losses (default: -0.1). If NULL, segments are not classified.}

\item{gain.lrr}{numeric. Cutoff value of lrr for gains (default: 0.1). If NULL, segments are not classified.}
}
\value{
A ggplot object
}
\description{
Plot copy ratio data
}
