% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_segmentation.R
\name{run_segmentation}
\alias{run_segmentation}
\title{Run segmentation using copynumber}
\usage{
run_segmentation(
  probeData,
  denoise.method = c("none", "winsorize", "rmf"),
  kmin = 5,
  gamma = 40,
  normalize = TRUE,
  k = 25,
  tau = 1.5,
  adjust.baseline = FALSE,
  use.n.probes = TRUE,
  n.cores = 1L,
  verbose = FALSE
)
}
\arguments{
\item{probeData}{tbl. seqnames (chromosomes chr1, ..., chrX), start, end, chrArm (chr1p, ..., chrXq), lrr (log R ratio).}

\item{denoise.method}{character. none, winsorize or rmf (recursive median filter).}

\item{kmin}{integer. minimum number of probes (or bins) in each segment used by copynumber::pcf (default: 5).}

\item{gamma}{numeric. A penalty parameter used by copynumber::pcf for each discontinuity in the curve (default: 40).}

\item{normalize}{logical. Whether the copy number measurements should be scaled by the sample residual standard error (default: TRUE).
When denoise.method = "winsorize", normalize = TRUE is recommended while when denoise.method = "rmf", normalize = FALSE is recommended.}

\item{k}{integer. window size to be used for the sliding window (actually half-window size) for copynumber::winsorize or denoise_intensities.}

\item{tau}{numeric. Winsorization threshold, default is 1.5.}

\item{adjust.baseline}{logical. Shift factor is calculated using limma::weighted.median of lrr among potentially
copy neutral segments (abs(lrr) < 0.1) (default: FALSE).}

\item{use.n.probes}{logical. If TRUE, the number of probes (bins) will be used as weights in the baseline adjusment.
If FALSE, the physical length of segments will be used as weights. (Default: TRUE)}

\item{n.cores}{integer. number of cores to be used (default: 1), n.cores > 1 is effective only for \code{denoise.method = "rmv"}.}

\item{verbose}{logical.}
}
\value{
list.
}
\description{
Run segmentation using copynumber
}
