% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_segments.R
\name{smooth_segments}
\alias{smooth_segments}
\title{Smooth segments}
\usage{
smooth_segments(
  segs,
  loss.lrr = -0.1,
  gain.lrr = 0.1,
  use.pcf = FALSE,
  gamma = 40,
  max.dist = 0.1
)
}
\arguments{
\item{segs}{list. Output of run_segmentation or annotate_segments.}

\item{loss.lrr}{numeric. Cutoff value of lrr for losses (default: -0.1). If NULL, segments are not classified.}

\item{gain.lrr}{numeric. Cutoff value of lrr for gains (default: 0.1). If NULL, segments are not classified.}

\item{use.pcf}{logical. Use PCF method for smoothing within each CNV status (default: FALSE).}

\item{gamma}{numeric. penalty for each discontinuity in the curve (default: 40).}

\item{max.dist}{numeric. the maximu difference of mean.lrr between the neighboring segments to be merged (default: 0.1).}
}
\value{
list. The same list with additional tbl named smoothed.
}
\description{
Merge consecutive segments in the same copy number state (loss, copy-neutral or gain).
}
