% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_segments.R
\name{annotate_segments}
\alias{annotate_segments}
\title{Annotate segments}
\usage{
annotate_segments(segs)
}
\arguments{
\item{segs}{list. Output of run_segmentation.}
}
\value{
list. The same list with additional columns in segs$segments.
}
\description{
Annotate segments
}
