% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pon.R
\name{make_pon}
\alias{make_pon}
\title{Make a panel of normals from SigDF data}
\usage{
make_pon(
  sdfs,
  platform = c("HM450", "EPIC", "EPICv2"),
  cutoff = -3.5,
  use.mask = TRUE,
  max.p.missing.sample = 0.2,
  percentile.var = 1,
  n.cores = 1L,
  median.normalize = TRUE,
  exclude.acen = FALSE,
  exclude.regions = NULL
)
}
\arguments{
\item{sdfs}{list. A named list of SigDF data.}

\item{platform}{character. HM450, EPIC, or EPICv2.}

\item{cutoff}{a cutoff value of log2(medianY/medianX) for sex inference (default: -3.5).}

\item{use.mask}{logical. Use SeSAMe mask to filter signals (default: TRUE).}

\item{max.p.missing.sample}{numeric. Maximum proportion of missing data per sample (default: 0.2).}

\item{percentile.var}{numeric. Remove top \code{percentile.var} percentile of highest variance (default: 1 percentile).}

\item{n.cores}{integer. Number of cores to be used to process data.}

\item{median.normalize}{logical. If TRUE, divide the intensity by their median (default: TRUE).}

\item{exclude.acen}{logical. If TRUE, probes within gieStain = "acen" will be removed (default: FALSE).}

\item{exclude.regions}{GenomicRanges object. Regions to be removed from the analysis (default: NULL).
A predefined set is available: data(exclude_regions.hg38).}
}
\value{
list. probeCoords, GenomicRanges object for probe coordinates. pred.sex, predicted sex.
data, a matrix of total intensities that were divided by sample median and log2 transformed.
}
\description{
Make a panel of normals from SigDF data
}
