% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pon.R
\name{make_pon}
\alias{make_pon}
\title{Make a panel of normals from SigDF data}
\usage{
make_pon(sdfs, max.p.missing = 0.01, percentile.var = 1)
}
\arguments{
\item{sdfs}{list. A named list of SigDF data.}

\item{max.p.missing}{numeric. Maximum proportion of missing data per probe (default: 0.01).}

\item{percentile.var}{numeric. Remove top \code{percentile.var} percentile of highest variance (default: 1 percentile).}
}
\value{
list. probeCoords, GenomicRanges object for probe coordinates. data, total intensities that
were divided by sample median and log2 transformed
}
\description{
Make a panel of normals from SigDF data
}
