% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangent_normalization.R
\name{tangent_normalization}
\alias{tangent_normalization}
\title{Tangent normalization of probe intensities}
\usage{
tangent_normalization(
  sdf,
  sex = NULL,
  cutoff = -3.5,
  pon,
  use.mask = TRUE,
  winsorize = FALSE,
  k = 25,
  tau = 1.5
)
}
\arguments{
\item{sdf}{SigDF object.}

\item{sex}{character. Sex of the sample. If NULL, sex is estimated using \code{get_sex_info}.}

\item{cutoff}{a cutoff value of log2(medianY/medianX) for sex inference (default: -3.5).}

\item{pon}{list. panel of normal data.}

\item{use.mask}{logical. Use SeSAMe mask to filter signals (default: TRUE).}

\item{winsorize}{logical. Winsorize after normalization (default: FALSE).}

\item{k}{integer. window size to be used for the sliding window (actually half-window size) for copynumber::winsorize.}

\item{tau}{numeric. Winsorization threshold, default is 1.5.}

\item{platform}{character. HM450, EPIC, or EPICv2.}
}
\value{
list. probeCoords, probe coordinates. obs, observed signal (\code{sesame::totalIntensities(sdf)}).
pred, predicted (fitted) signal.
lrr, log R ratio (\code{obs - pred}). shift, shift factor of the baseline signal.
}
\description{
Tangent normalization of probe intensities
}
