% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tangent_normalization.R
\name{tangent_normalization}
\alias{tangent_normalization}
\title{Tangent normalization of probe intensities}
\usage{
tangent_normalization(sdf, sex = NULL, pon)
}
\arguments{
\item{sdf}{SigDF object.}

\item{sex}{character. Sex of the sample. If NULL, sex is estimated using \code{sesame::inferSex(sdf)}.}

\item{pon}{list. panel of normal data.}
}
\value{
list. probeCoords, probe coordinates. obs, observed signal (\code{sesame::totalIntensities(sdf)}). pred, predicted (fitted) signal.
lrr, log R ratio (\code{obs - pred}). shift, shift factor of the baseline signal.
}
\description{
Tangent normalization of probe intensities
}
